Remove old, confusingly-named "master" branch
This branch was intended to feed the packaging effort by tracking the
upstream development master in git, hence the (rather confusing) name.
However it never really got used for anything and has rotted a bit, so
it's best abandoned I think.

Upstream maintains (okay, I maintain) an upstream debianization now,
which will serve the same purpose and allow us to maintain a PPA for
testers, which I want to do.

https://github.com/mypaint/debian
https://github.com/mypaint/mypaint/issues/191

The Debian trunk/ is unaffected by this, and will be updated in due
course (for the next release, or for decent ~rc versions), hopefully
adopting some or all of the Debian-style packaging work happening
upstream in the form of patches rather than from a shared history.

D    branches/master